+++ mobile/android/android-components/components/support/locale/src/test/java/mozilla/components/support/locale/LocaleManagerTest.kt
@@ -5,22 +5,32 @@
 package mozilla.components.support.locale
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.testContext
+import mozilla.components.support.test.whenever
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotEquals
 import org.junit.Assert.assertNull
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mockito.Mockito.times
+import org.mockito.Mockito.verify
+import org.mockito.Mockito.never
 import org.robolectric.annotation.Config
 import java.util.Locale
 
 @RunWith(AndroidJUnit4::class)
 class LocaleManagerTest {
 
+    private lateinit var localeUseCases: LocaleUseCases
+
     @Before
     fun setup() {
         LocaleManager.clear(testContext)
+
+        localeUseCases = mock()
+        whenever(localeUseCases.notifyLocaleChanged).thenReturn(mock())
     }
 
     @Test
@@ -30,7 +40,7 @@ class LocaleManagerTest {
 
         assertNull(currentLocale)
 
-        val newContext = LocaleManager.setNewLocale(testContext, "es")
+        val newContext = LocaleManager.setNewLocale(testContext, localeUseCases, "es".toLocale())
 
         assertNotEquals(testContext, newContext)
 
@@ -54,13 +64,18 @@ class LocaleManagerTest {
     fun `when resetting to system locale then the current locale will be the system one`() {
         assertEquals("en_US".toLocale(), Locale.getDefault())
 
-        LocaleManager.setNewLocale(testContext, "fr")
+        LocaleManager.setNewLocale(
+            testContext,
+            localeUseCases,
+            "fr".toLocale()
+        )
+
         val storedLocale = Locale.getDefault()
 
         assertEquals("fr".toLocale(), Locale.getDefault())
         assertEquals("fr".toLocale(), storedLocale)
 
-        LocaleManager.resetToSystemDefault(testContext)
+        LocaleManager.resetToSystemDefault(testContext, localeUseCases)
 
         assertEquals("en_US".toLocale(), Locale.getDefault())
         assertNull(LocaleManager.getCurrentLocale(testContext))
@@ -71,10 +86,68 @@ class LocaleManagerTest {
     fun `when setting a new locale then the current locale will be different than the system locale`() {
         assertEquals("en_US".toLocale(), Locale.getDefault())
 
-        LocaleManager.setNewLocale(testContext, "fr")
+        LocaleManager.setNewLocale(testContext, localeUseCases, "fr".toLocale())
 
         assertEquals("en_US".toLocale(), LocaleManager.getSystemDefault())
         assertEquals("fr".toLocale(), LocaleManager.getCurrentLocale(testContext))
         assertEquals("fr".toLocale(), Locale.getDefault())
+    }
+
+    @Test
+    @Config(qualifiers = "en-rUS")
+    fun `when setting a new locale then the store is notified via use case`() {
+        assertEquals("en_US".toLocale(), Locale.getDefault())
+
+        LocaleManager.setNewLocale(testContext, localeUseCases, "fr".toLocale())
+
+        verify(localeUseCases, times(1)).notifyLocaleChanged
+    }
+
+    @Test
+    @Config(qualifiers = "en-rUS")
+    fun `GIVEN the locale use cases are defined WHEN resetting to system default locale THEN the store is notified via use case`() {
+        assertEquals("en_US".toLocale(), Locale.getDefault())
+
+        LocaleManager.setNewLocale(
+            testContext,
+            localeUseCases,
+            "fr".toLocale()
+        )
+
+        verify(localeUseCases, times(1)).notifyLocaleChanged
+
+        val storedLocale = Locale.getDefault()
+
+        assertEquals("fr".toLocale(), Locale.getDefault())
+        assertEquals("fr".toLocale(), storedLocale)
+
+        LocaleManager.resetToSystemDefault(testContext, localeUseCases)
+
+        verify(localeUseCases, times(2)).notifyLocaleChanged
+    }
+
+    @Test
+    @Config(qualifiers = "en-rUS")
+    fun `WHEN setting a new locale THEN locale use cases are used only when defined`() {
+        assertEquals("en_US".toLocale(), Locale.getDefault())
+
+        val locale1 = "fr".toLocale()
+        LocaleManager.setNewLocale(
+            testContext,
+            locale = locale1
+        )
+
+        verify(localeUseCases, never()).notifyLocaleChanged
+        assertEquals(locale1, Locale.getDefault())
+
+        val locale2 = "es".toLocale()
+        LocaleManager.setNewLocale(
+            testContext,
+            localeUseCases,
+            locale2
+        )
+
+        verify(localeUseCases, times(1)).notifyLocaleChanged
+        assertEquals(locale2, Locale.getDefault())
     }
 }